Reworked the tests
Okay, I did quite a few iterations here and this is what I landed one:

- We should not pass the test object into any factories. That hides
what is actually being tested. It is risky in terms of wrong logic
inside factory. Some of the tests still ran green after changing to
something that should not, also after my initial iterations

- The test object should test each route separately. This will ensure
any failing tests to give out the correct line number and we can test
other stuff if we want to

- I created one smaller factory that just creates the controller with a
single signal for testing route matching and the router itself

- We can test routing by matching on Url or running a signal manually

- So now it tests a route change by manually changing the url of
addressbar and by running a signal manually. No tests related to
addressbar emitting an event as that just results in setting the value
of addressbar

Changing any of the tests now results in bad tests :-)